break predict et al into separate file
