+++ src/main.cpp
@@ -37,6 +37,7 @@ PYBIND11_MODULE(_core, m)
     orbit_fit::gauss_bindings(m);
     orbit_fit::orbit_fit_bindings(m);
     orbit_fit::orbit_fit_result_bindings(m);
+    orbit_fit::predict_bindings(m);
     orbit_fit::predict_result_bindings(m);
 
 #ifdef VERSION_INFO
